requirement_haici.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argbind==0.3.7
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audiomentations==0.34.1
audioread==3.0.1
Babel==2.14.0
backoff==2.2.1
beartype==0.17.2
beautifulsoup4==4.12.3
bibtexparser==2.0.0b7
blessed==1.20.0
boto3==1.34.45
botocore==1.34.45
braceexpand==0.1.7
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.1
cmake==3.28.3
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.0
croniter==1.4.1
csvw==3.3.0
cycler==0.12.1
Cython==3.0.8
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.7.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
dill==0.3.8
docker-pycreds==0.4.0
docstring-parser==0.15
editor==1.6.6
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
fire==0.5.0
flatten-dict==0.4.2
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.0
google-auth-oauthlib==1.0.0
gradio==4.19.1
gradio_client==0.10.0
grpcio==1.60.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.6
importlib-resources==6.1.1
inquirer==3.2.3
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.3
librosa==0.10.1
lightning==2.0.9.post0
lightning-cloud==0.5.64
lightning-utilities==0.10.1
lightning-vocoders @ git+https://github.com/Wataru-Nakata/ssl-vocoders@87ee8a4239c587b10b04c8b310684d35a9c64d25
lit==17.0.6
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
orjson==3.9.14
packaging==23.2
pandarallel==1.6.5
pandas==2.2.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pylatexenc==2.10
pyloudnorm==0.1.1
pyparsing==3.1.1
pyroomacoustics==0.7.3
pyrootutils==1.0.4
pysndfx==0.3.6
pystoi==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
randomname==0.2.1
rdflib==7.0.0
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.0
rotary-embedding-torch==0.5.3
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.2.2
runs==1.2.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segments==2.2.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.40.5
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==0.5.16
stack-data==0.6.3
starlette==0.36.3
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
text2phonemesequence==0.1.4
threadpoolctl==3.3.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch-stoi==0.2.1
torchaudio==2.0.2
torchmetrics==1.0.1
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
triton==2.0.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.27.1
wandb==0.15.12
wcwidth==0.2.13
webdataset==0.2.86
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
xmod==1.8.1
xvector_jtubespeech==0.0.2
yarl==1.9.4
